requirements.txt
streamlit
groq
edge-tts
pexels-api
moviepy<2.0.0
python-dotenv
asyncio
requests
